move theme to component
