+++ pages/page-layout-examples/FirebaseDemo/FirebaseDemo.js
@@ -1,13 +1,55 @@
 import React from 'react';
 import MuiThemeProvider from '@material-ui/core/styles/MuiThemeProvider';
-import withStyles from '@material-ui/core/styles/withStyles';
+import { createMuiTheme, withStyles } from '@material-ui/core/styles';
 // COMPONENTS
 import CssBaseline from '@material-ui/core/CssBaseline';
 import Navigator from './components/Navigator';
 import Content from './components/Content';
 import Header from './components/Header';
 // THEME
-import theme from './theme';
+
+const theme = createMuiTheme({
+  typography: {
+    useNextVariants: true,
+  },
+  palette: {
+    primary: {
+      light: '#63ccff',
+      main: '#009be5',
+      dark: '#006db3',
+      contrastText: '#fff',
+    },
+  },
+  shape: {
+    borderRadius: 8,
+  },
+  overrides: {
+    MuiDrawer: {
+      paper: {
+        background: '#18202c',
+      },
+    },
+    MuiButton: {
+      root: {},
+      label: {
+        textTransform: 'initial',
+      },
+      contained: {
+        boxShadow: 'none',
+      },
+    },
+    MuiPrivateTabIndicator: {
+      root: {
+        height: 3,
+        borderTopLeftRadius: 3,
+        borderTopRightRadius: 3,
+      },
+      colorSecondary: {
+        backgroundColor: '#fff',
+      },
+    },
+  },
+});
 
 const drawerWidth = 256;
